refactor: align the PartialEq impl for unsigned integers with the one for signed integers
